Fix cross-platform bugs found in iOS/Android audit
Android:
- Widen prize-pool money math to Long to avoid Int32 overflow on large
  buy-in × entry counts (matches iOS 64-bit); Payout.amount now Long
- TournamentState default remainingSeconds now honors the passed config
  instead of a fresh default config
- Keep screen awake while the dealer turn timer runs, not just the clock
- LogoStorage: guard createScaledBitmap against a 0 dimension for extreme
  aspect ratios (IllegalArgumentException)
- NumberField: local text buffer so fields can be cleared/retyped and
  large input can't silently collapse to 0 (9-digit cap keeps Int range)

iOS:
- TournamentStore: mutate @Observable state on the main actor (ticker Task
  now @MainActor) to fix an off-main data race with pause/reset
- TurnTimerEngine: invalidate the repeating Timer in deinit (leak)
- Keep screen awake while the dealer turn timer runs
- ContentView: bind TabView selection so changing theme no longer snaps
  the user off their current tab
- SettingsView: reset selectedPhoto after import so re-picking the same
  photo works
- TurnTimerView: re-sync the time-chip field on slot switch to avoid a
  stale value being written back

Both platforms build clean (debug + release AAB on Android, iOS simulator).
Reviewed by parallel audit agents and adversarially re-verified.

+++ android/app/src/main/java/com/jiny/raisetimer/domain/PayoutCalculator.kt
@@ -4,7 +4,7 @@ import com.jiny.raisetimer.domain.model.TournamentState
 
 object PayoutCalculator {
 
-    data class Payout(val place: Int, val amount: Int)
+    data class Payout(val place: Int, val amount: Long)
 
     /**
      * Splits the total prize pool across [TournamentConfig.payoutPercents]. Percentages are
@@ -16,10 +16,10 @@ object PayoutCalculator {
         if (percents.isEmpty()) return emptyList()
 
         val pool = state.totalPrizePool
-        if (pool <= 0) return percents.indices.map { Payout(it + 1, 0) }
+        if (pool <= 0) return percents.indices.map { Payout(it + 1, 0L) }
 
         val sum = percents.sum().takeIf { it > 0 } ?: 100
-        val rawAmounts = percents.map { pct -> (pool.toLong() * pct / sum).toInt() }
+        val rawAmounts = percents.map { pct -> pool * pct / sum }
         val distributed = rawAmounts.sum()
         val remainder = pool - distributed
